R3 fix: validate host-supplied enum discriminants (uint32_t + range-check)
The M7.7 review found the 4 host-supplied #[repr(u32)] enum params (DoodleDirective,
DoodleObservationMode, DoodleBodyKind, DoodleBuiltin) were received by value, so an
out-of-range discriminant was instant UB at the boundary (reachable via version
skew or a host bug) that the panic firewall cannot catch -- violating "never UB
across the boundary". User ratified: validate.

Each param now crosses as uint32_t and is range-checked to a known variant,
returning ErrContract (doodle_drive/_slice, doodle_set_observation_mode,
doodle_registry_add_builtin), NULL (doodle_foreign_desc_new -- its failure signal),
or ignored (doodle_config_set_observation_mode -- a void setter). This completes
D-M7-3's symmetry: engine->host unknowns cross via the reserved unknown-tag,
host->engine unknowns validate. Added freeze convention 7 (lib.rs) so every future
host-supplied enum param follows suit. The named enum types stay in the header
(forced via cbindgen [export].include) so hosts keep the constants; params doc-name
their enum. Also wrapped doodle_registry_add_builtin in guard::catch (a prior
convention-5 gap). Test: an out-of-range value to each param is a defined error,
not UB (now expressible -- passes under Miri). Regenerated doodle.h.

+++ crates/doodle-capi/tests/gc_stress.rs
@@ -74,7 +74,7 @@ fn a_foreign_value_finalizes_once_at_gc_under_stress_through_c() {
     // reclaimed *during the drive* — its finalizer runs at GC time, through the C trampoline.
     let mut out = DoodleOutcome::blank();
     assert_eq!(
-        unsafe { doodle_drive(inst, DoodleDirective::RunToCompletion, &mut out) },
+        unsafe { doodle_drive(inst, DoodleDirective::RunToCompletion as u32, &mut out) },
         DoodleStatus::Ok
     );
     assert_eq!(out.kind, DoodleOutcomeKind::Completed);
